% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_FCFF.R
\name{Calculate_FCFF}
\alias{Calculate_FCFF}
\title{Calculate_FCFF}
\usage{
Calculate_FCFF(Ticker, year)
}
\arguments{
\item{Ticker}{Ticker of the stock to be evaluated}

\item{year}{Year to calculate FCFF for}
}
\value{
FCFF Returns the calculated FCFF
}
\description{
The function calculates the free cash flow to firm (FCFF) using the cash flow financials from GetCashFlow function and the Morningstar data.
             The FCFF is calculated from the Net Income, Change in working capital, depreciation expense and Capital expenditures (CAPEX).
}
\examples{
\dontrun {

Ticker <- "AAPL"
year <- 2017
FCFF_AAPL_2017 <- Calculate_FCFF(Ticker, year)

}
}
\seealso{
\code{\link[finreportr]{GetCashFlow()}}
}
