% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCFE_Basic.R
\name{FCFE_Basic}
\alias{FCFE_Basic}
\title{FCFE_Basic}
\usage{
FCFE_Basic(Ticker, Req, year, growth_rate)
}
\arguments{
\item{Ticker}{Ticker of the company to be evaluated}

\item{Req}{Required rate of return}

\item{year}{Year when latest results is available}

\item{growth_rate}{The growth rate g}
}
\value{
FCFE_Value
}
\description{
Basic valuation of a company based on the FCFE. The FCFE is calculated by adjusting the FCFF by Net Borrowings
}
