% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCFF_Basic.R
\name{FCFF_Basic}
\alias{FCFF_Basic}
\title{FCFF_Basic}
\usage{
FCFF_Basic(Ticker, WACC, growth_rate)
}
\arguments{
\item{Ticker}{The ticker of the stock to be evaluated}

\item{WACC}{The WACC as a percentage}

\item{growth_rate}{The growth rate g}
}
\value{
FCFF_Value
}
\description{
Basic valuation of a company using the free cash flow to firm. Value = (FCFF(1+g))/(WACC - g)
}
