% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gordon_Growth.R
\name{Gordon_Growth}
\alias{Gordon_Growth}
\title{Gordon_Growth}
\usage{
Gordon_Growth(Ticker, Req_Rate)
}
\arguments{
\item{Ticker}{Stock to be evaluated}

\item{Req_Rate}{Required rate of return for specific stock}
}
\value{
Gordon_g The growth rate output
}
\description{
The function calculates the growth rate based on the Gordon Growth model: g = r - D1/P
}
\examples{
\dontrun {

 Gordon_Growth("AAPL", Req_Rate = 0.2)

}
}
\seealso{
\code{\link[tidyquant]{tq_get}}
}
