% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Justified_PB.R
\name{Justified_PB}
\alias{Justified_PB}
\title{Justified_PB}
\usage{
Justified_PB(Ticker, Req, growth_rate = 0)
}
\arguments{
\item{Ticker}{The symbol of the stock to be evaluated}

\item{Req}{The required return on equity}

\item{growth_rate}{Optional growth_rate as percentage, otherwise calculated using Growth_Rate function}
}
\value{
Just_PB The justified PB output
}
\description{
Calculate the justified Price to Book ratio based on fundamentals
}
