% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Justified_PE.R
\name{Justified_PE}
\alias{Justified_PE}
\title{Justified_PE}
\usage{
Justified_PE(Ticker, Req, growth_rate = 0)
}
\arguments{
\item{Ticker}{Symbol of stock to be evaluated}

\item{Req}{The Required return on equity}

\item{growth_rate}{Optional growth_rate (%), otherwise calculated using Growth_Rate function}
}
\value{
Just_PE The justified PE as output
}
\description{
Calculate the justified PE's
}
