% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Justified_PS.R
\name{Justified_PS}
\alias{Justified_PS}
\title{Justified_PS}
\usage{
Justified_PS(Ticker, Req, growth_rate = 0)
}
\arguments{
\item{Ticker}{The symbol of the stock to be evaluated}

\item{Req}{The required return on equity}

\item{growth_rate}{Optional growth_rate as percentage, otherwise calculated using Growth_Rate function}
}
\value{
Just_PS The justified PS output
}
\description{
Calculate the justified Price to Sales using fundamentals
}
