% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_DDM.R
\name{Multi_DDM}
\alias{Multi_DDM}
\title{Multi_DDM}
\usage{
Multi_DDM(Ticker, Req, growth_rate, terminal_g, length = 5)
}
\arguments{
\item{Ticker}{Provide the stock ticker, which will be searched for on morningstar and yahoo finance}

\item{Req}{The required rate of return used for discounting}

\item{growth_rate}{growth_rate now}

\item{terminal_g}{The terminal growth rate}

\item{length}{The discounted length from first to terminal dividend, default is 5}
}
\value{
DDM_Out The results of the DDM model
}
\description{
The function calculates stock value based on a single stage DDM model (Gordon Growth Model)
}
\examples{
{

 Ticker <- "MSFT"
 Req <- 0.12
 growth_rate <- 0.11
 terminal_g <- 0.08
 length <- 5
 Multi_Stage_DDM <- Multi_DDM(Ticker, Req, growth_rate, terminal_g, length)


}
}
\seealso{
\code{\link[tidyquant]{tq_get}}
}
