% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Ratios.R
\name{Plot_Ratios}
\alias{Plot_Ratios}
\title{Plot_Ratios}
\usage{
Plot_Ratios(Ticker, Ratios)
}
\arguments{
\item{Ticker}{The stock Ticker to be evaluated}

\item{Ratios}{The ratios you wish to plot. Can be either a single charactar or character vector.}
}
\value{
Out_plots The output containing all plots as a list
}
\description{
Plot financial ratios using the morningstar data of a specific company.
}
\examples{
{

 Plot_Ratios("AAPL", "Net Income USD")

}
}
\seealso{
\code{\link[tidyquant]{tq_get}}
}
