% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price_Momentum.R
\name{Price_Momentum}
\alias{Price_Momentum}
\title{Price_Momentum}
\usage{
Price_Momentum(Ticker, Times)
}
\arguments{
\item{Ticker}{Stock ticker to be evaluated}

\item{Times}{List of time periods in months}
}
\value{
Mom Momentum output
}
\description{
Get price momentum for given time periods
}
\examples{
\dontrun {

}
}
\seealso{
\code{\link[tidyquant]{tq_get}}
}
