% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_DDM.R
\name{Single_DDM}
\alias{Single_DDM}
\title{Single_DDM}
\usage{
Single_DDM(Ticker, Req, growth_rate)
}
\arguments{
\item{Ticker}{Provide the stock ticker, which will be searched for on morningstar and yahoo finance}

\item{Req}{The required rate of return used for discounting as a fraction}

\item{growth_rate}{growth_rate}
}
\value{
DDM_Out The results of the DDM model
}
\description{
The function calculates stock value based on a single stage DDM model (Gordon Growth Model)
}
\examples{
 {

#Ticker <- "MSFT"
#Req <- 0.12
#growth_rate <- 0.1
 # Single_Stage_DDM <- Single_DDM(Ticker, Req, growth_rate)

}
}
\seealso{
\code{\link[tidyquant]{tq_get}}
}
