% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wacc.R
\name{WACC}
\alias{WACC}
\title{WACC}
\usage{
WACC(Ticker, cost_Equity, cost_Debt, cost_Pref, Tax_Rate, Per_Equity = 0,
  Per_Debt = 0)
}
\arguments{
\item{Ticker}{The ticker of the company to evaluated}

\item{cost_Equity}{The cost of equity as a fraction}

\item{cost_Debt}{The cost of debt as a fraction}

\item{cost_Pref}{The cost of Preferred stock as a fraction}

\item{Tax_Rate}{The tax rate (%)}

\item{Per_Debt}{Optional, the percentage debt}

\item{Per_Equit}{Optional, the percentage equity}
}
\value{
WACC Return the WACC object
}
\description{
Calculate a company's WACC based on cost of Equity and cost of debt. The fraction of debt and equity can be provided by the user.
 Otherwise, the function aims to calculate the weights based on most recent balance sheet data collected from the getBalancesheet() function
}
